backend/requirements.txt
Django>=4.0
djangorestframework
psycopg2-binary
gunicorn
drf-yasg
environs